requirements.txt
apispec==6.8.4
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-smorest==0.46.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.0.1
mysql-connector-python==9.5.0
packaging==25.0
webargs==8.7.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
PyJwt==2.10.0
Flask-JWT-Extended==4.7.1
Flask-Mail
requests
Flask-SQLAlchemy==3.1.1